requirements.txt
# Django Web Framework
Django==5.0.6

# Django Cors Headers - Used to enable CORS headers in API responses, and allow requests to be made to your API server from other origins.
django-cors-headers==4.3.1

# Django Rest Framework - Used to design API logic
djangorestframework==3.15.0

# MySQL Client - Used as an interface to connect Django application to the MySQL server.
#mysqlclient==2.0.1